+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 # Changelog
 
-All notable changes to `robtrehy/laravel-user-permissions` will be documented in this file.
+All notable changes to `robtrehy/laravel-user-preferences` will be documented in this file.
+
+## 4.1.0 - 2025-11-09
+- Added methods for working with preferences for any user instance or ID:
+  - `getForUser()`
+  - `setForUser()`
+  - `resetForUser()`
+  - `hasForUser()`
+- Achieved 100% test coverage for `UserPreferences` class.
+- Updated internal caching to support per-user caching (`$preferencesCache` and `$hasLoadedCache`).
+- Minor docblock improvements.
+
+## 4.0.0 - 2025-08-01
+- Added support for Laravel 12
+- Removed support for Laravel 9
 
 ## 3.1.0 - 2024-03-18
 - Added support for Laravel 11